Use a PSR-3 logger for emitting messages
Per a comment on #5, this patch modifies the `RequestHandlerSwooleRunner`
to compose a PSR-3 logger and use it to emit any messages to the process
invoking the server.

The logger is optional. If none is provided, a default implementation,
`StdoutLogger`, is provided. This implementation ignores the log level,
but interpolates any variables in the `$context` into the `$message`
before calling `printf()` to emit the message with a trailing `PHP_EOL`
character.