Remove dependency on zend-console
This patch removes the dependency on zend-console. It does so by adding
two classes, `CommandLine` and `CommandLineOptions`. The first is
responsible for parsing the CLI arguments and options, and the second is
a value object with the results. In the situation that it cannot parse
the command line (unknown options or arguments), it will emit an error
message and the help message. When `--help`, `-h`, or `help` are part of
the command line, it will emit the help message.

The code is based lightly on how zend-console's `Getopt` class works,
but streamlined for the expected options and arguments of this package.