Extract static resource handling, and allow varying headers by path
This patch refactors the static resource handling. First, it creates a
new interface, `StaticResourceHandlerInterface`; instances of this
interface are optionally injected into `RequestHandlerSwooleRunner`
instances via the constructor.

When present, the `RequestHandlerSwooleRunner` will first call its
`isStaticResource()` method, and, if true, its `sendStaticResource()`
method during `onRequest()`.

`StaticResourceHandler` now provides the default implementation. It
receives:

- The document root
- An array of extension => mime type pairs. `null` indicates the
  defaults should be used, while an empty array effectively disables
  static resource handling.
- Cache-Control rules: match the regexp to apply one or more directives
- Last-Modified rules: match the regexp to enable
- ETag rules: match the regexp to enable

As before, it checks to see if a request URI matches any of the
extensions in the type map; if not, it is not considered a static
resource. It then checks for cache-control directives matching the
current path, a last-modified flag matching the current path, and an
ETag flag matching the current path, emitting appropriate headers in
each case, and, in the case of Last-Modified and ETag, potentially
sending an empty response with a 304 header when appropriate conditions
are met.

Additionally, this patch modifies the workflow to accommodate the
following:

- Emitting 405 responses with `Allow` headers when the file exists, but
  an invalid method is issued.
- Handling OPTIONS requests to known files, emitting a 200 response with
  an `Allow` header and all relevant caching headers.

A new `StaticResourceHandlerFactory` accepts configuration in the
following format:

```php
'zend-expressive-swoole' => [
    'swoole-http-server' => [
        'document-root' => '/path/to/static/files/to/serve', // usu getcwd() . /public/
        'type-map' => [
            // extension => mimetype pairs of types to cache.
            // A default list exists if none is provided.
        ],
        'etag-type' => 'weak|strong', // ETag algorithm type to use, if any
        'directives' => [
            // Rules governing which server-side caching headers are emitted.
            // Each key must be a valid regular expression, and should match
            // typically only file extensions, but potentially full paths.
            // When a static resource matches, all associated rules will apply.
            'regex' => [
                'cache-control' => [
                    // one or more valid Cache-Control directives:
                    // - must-revalidate
                    // - no-cache
                    // - no-store
                    // - no-transform
                    // - public
                    // - private
                    // - max-age=\d+
                ],
                'last-modified' => bool, // Emit a Last-Modified header?
                'etag' => bool, // Emit an ETag header?
            ],
        ],
    ],
],
```