Use a middleware approach for static resources
I found the logic for handling static resources was rather unwieldy, as
it involved a ton of conditionals, particularly around whether or not to
send content back to the client.

This patch introduces a middleware approach to handling static
resources. The `StaticResourceHandler` continues to determine if a
static resource was matched, as well as to emit a `Content-Type` header
when returning a static resource. However, instead of having logic for
the various caching and error conditions, it dispatches a middleware
queue. This approach allows users to layer a set of middleware for
handling the static resource. I extracted the following middleware (all
under the `StaticResourceHandler` subnamespace):

- `MethodNotAllowedMiddleware`
- `OptionsMiddleware`
- `HeadMiddleware`
- `ClearStatCacheMiddleware`
- `CacheControlMiddleware`
- `LastModifiedMiddleware`
- `ETagMiddleware`

The `StaticResourceHandlerFactory` now conditionally adds the last four
middleware based on the presence of applicable configuration (the first
three are always added).

Each middleware receives the Swoole request and the filename of the
resource, and is expected to return a new `ResponseValues` instance.
This latter aggregates headers, the status, and a flag as to whether or
not to send content. The content flag defaults to send the file, but any
middleware at any time can disable sending content - but never the
opposite!

Additionally, this approach allows for more granular and simpler testing
of each static resource directive; the unit tests for each middleware
bear this out. Integration tests can verify that expectations are
correct when middleware is composed in  a specific order; the
`StaticResourceHandler\IntegrationTest` demonstrates this.