Refactor ResponseValues to StaticResourceResponse
Renames `ResponseValues` to `StaticResourceResponse`, and makes it
responsible for emitting the Swoole HTTP Response representing the
static resource. This will allow us to make the gzip/deflate
functionality its own middleware, as it can inject the appropriate
callback for emitting the response body.

This change means that the various getter methods are no longer
relevant, as the only object accessing them is itself.

To ensure we can still test against the `StaticResourceResponse` state,
this patch adds an `AssertResponseTrait`. This will also insulate us
from further changes.