Refactor gzip/deflate support to static resource middleware
This patch refactors the gzip/deflate support provided in #13 to be
provided as static resource middleware via the class
`Zend\Expressive\Swoole\StaticResourceHandler\GzipMiddleware`. The patch
adds tests for the new middleware, and integrates it into the default
pipeline created in the `StaticResourceHandlerFactory`. It is also now
present in the unit tests, though they do not test compressing any
content at this time.