Replace custom CLI parsing with symfony/console
This approach allows us to grow more easily. `CommandLine` and
`CommandLineOptions` were replaced with three commands in the
subnamespace `Zend\Expressive\Swoole\Command` (one for each of start,
stop, and reload), and these each receive the runner as an argument,
calling methods on it based on the action and options provided.

This also means some changes to the CLI structure:

- "-n" is taken by symfony/console to mean "--no-interaction". As such,
  `--num-workers` now has `-w` as its short form.
- While we can do a default command, default commands cannot accept
  arguments or options. As such, you MUST specify `start` now.

Overall, though, most changes are transparent to the end-user.