Fixes logging errors observed in integration tests
The method `getServerParam()` is expected to return a string, always,
but some data is presented as other scalar types. This patch now casts
it before returning it.

Additionally, I discovered `$request->get` is `null` when no query
string data is provided; this casts it to an array in such situations.