Ensure SSL support can be enabled
Per #44, while we allowed passing the SSL key and certificate as options
for the Swoole HTTP server, we were not allowing protocol options that
enabled SSL support. Specifically, you need to pass the union of
`SWOOLE_SSL` and one of either `SWOOLE_SOCK_TCP` or `SWOOLE_SOCK_TCP6`;
prior to this patch, the union types were not allowed.

This patch backports a test and test assets from #40, authored by
@gsteel, so that we can fix this in the version 1 series.