Move CLI tooling to dedicated binary
A problem we were observing is that if we want access to the HTTP server
for the purpose of registering task workers, triggering tasks, etc., we
MUST be able to compose the HTTP server directly in the HTTP runner.
However, if the commands are run via the HTTP runner, the server is
already present, and we cannot do things such as restart the server.

This patch thus moves the CLI tooling out of the runner, and creates a
vendor binary, `zend-expressive-swoole`. One nice aspect of this is that
`reload` now properly reloads the entire application, allowing it to
pick up code, configuration, and other changes.

I've also taken the liberty of adding a `status` command, to determine
whether or not the server is currently running.

The downside of this is that it cannot pick up changes to the workflow
of `public/index.php`. As such, if users want to change that, they will
need to do it in either `config/routes.php`, `config/pipeline.php`, or
in their own command line tooling.