Only set server options if they were provided to the command
Previously, a default value was used for the number of workers, and the
`daemonize` option toggled off if no options were passed. However, this
would override any _configuration_ options in that case -- which should
only happen if the option was passed to the command line.

As such, `StartCommand` now only pushes options passed on the command
line to the server, and only if any were provided.